requirements.txt
codetiming
matplotlib
morfeusz2
numpy
pandas
python-dotenv
regex
statsmodels
tqdm
